hdfs.py refactoring: Move out configuration
By putting it in a seperate file, it's easier to refactor the rest of
contents of that file.
